Purple: changed CN min/max to int from long

+++ hmf-common/src/main/java/com/hartwig/hmftools/common/purple/PurpleCopyNumber.java
@@ -41,9 +41,9 @@ public abstract class PurpleCopyNumber implements GenomeRegion {
         return end() - start() + 1;
     }
 
-    public abstract long minStart();
+    public abstract int minStart();
 
-    public abstract long maxStart();
+    public abstract int maxStart();
 
     public boolean svSupport() {
         return segmentStartSupport().isSV() || segmentEndSupport().isSV();
